Certificate expiry date check
This CloudFormation template sets up a Lambda function to monitor ACM certificate expiry dates at regular intervals and sends an SNS notification when it is about to expire. This is useful if you have imported certificates into ACM, but they are not renewed by ACM and you need to know if they are about to expire in good time so that replacements can be organised.